app/requirements.txt
asgiref==3.3.1
Django==3.1.6
gunicorn==20.0.4
loguru==0.5.3
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
django-cors-headers==3.7.0
djangorestframework==3.12.3
coverage==5.5
pyyaml==5.4.1
uritemplate==3.0.1
phonenumbers==8.12.20
django-phonenumber-field==5.0.0
